We want people to stop using email so having an email link in this modal is not right. We need to get communications flowing through the platform. For now, just remove it.
Remove licensed in from Contact modal.

Due to attorney-client privilege, Yael should not be part of the team as that implies she has access to the client's documents (which she doesn't unless they explicitly ask her). Please take her out of this view for now until we resolve how to best communicate her role.

Please change Customer to Client. Customer in legal contexts is weird. We only use it in this context because ross said that Client was a reserved word.